Atualizações no model e inclusão da classe Member (issue #4)

+++ src/main/java/br/ufrn/dct/apf/model/User.java
@@ -30,10 +30,10 @@ public class User implements Serializable {
      */
     private static final long serialVersionUID = 1L;
 
- @Id
+    @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "user_id")
-    private int id;
+    private Long id;
 
     @Column(name = "email")
     @Email(message = "*Please provide a valid Email")
@@ -61,11 +61,11 @@ public class User implements Serializable {
     @JoinTable(name = "user_role", joinColumns = @JoinColumn(name = "user_id"), inverseJoinColumns = @JoinColumn(name = "role_id"))
     private Set<Role> roles;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
